Move collections to the model classes
