+++ Jobber.Sdk/Models/Jobs/QuotesCollection.cs
@@ -1,12 +1,10 @@
 ﻿using System.Collections.Generic;
-using Jobber.Sdk.Models;
-using Newtonsoft.Json;
 using System.Linq;
-using Jobber.Sdk.Models.Jobs;
+using Newtonsoft.Json;
 
-namespace Jobber.Sdk.Responses
+namespace Jobber.Sdk.Models.Jobs
 {
-    public class QuotesResponse
+    public class QuotesCollection
     {
         [JsonProperty("quotes")]
         public IEnumerable<Quote> Quotes { get; set; } = new List<Quote>();
