+++ Jobber.Sdk/Models/Jobs/VisitsCollections.cs
@@ -1,12 +1,10 @@
 ﻿using System.Collections.Generic;
-using Jobber.Sdk.Models;
-using Newtonsoft.Json;
 using System.Linq;
-using Jobber.Sdk.Models.Jobs;
+using Newtonsoft.Json;
 
-namespace Jobber.Sdk.Responses
+namespace Jobber.Sdk.Models.Jobs
 {
-    public class VisitsResponse
+    public class VisitsCollections
     {
         [JsonProperty("visits")]
         public IEnumerable<Visit> Visits { get; set; } = new List<Visit>();
