+++ Jobber.SmartAssistant/Features/ConvertableQuotes/ConvertibleQuoteIntentFulfiller.cs
@@ -3,7 +3,6 @@ using DialogFlow.Sdk.Builders;
 using DialogFlow.Sdk.Models.Fulfillment;
 using Jobber.Sdk;
 using Jobber.Sdk.Models;
-using Jobber.Sdk.Responses;
 using Jobber.SmartAssistant.Core;
 
 namespace Jobber.SmartAssistant.Features.ConvertableQuotes
