+++ Jobber.SmartAssistant/Features/CreateJob/ClientRequestedCreateJobIntentFulfiller.cs
@@ -5,7 +5,6 @@ using DialogFlow.Sdk.Models.Fulfillment;
 using Jobber.Sdk;
 using Jobber.Sdk.Models;
 using Jobber.Sdk.Models.Clients;
-using Jobber.Sdk.Responses;
 using Jobber.SmartAssistant.Core;
 
 namespace Jobber.SmartAssistant.Features.CreateJob
